translate2csv/requirements.txt
py-translate
